Refactor gotls probe to follow standardized architecture pattern (#916)
## PR Merge 描述（提炼总结）

本 PR 以 **GoTLS 探针**为样板，完成从旧实现向新 `internal/probe` 架构的深度重构与对齐（对标 bash probe 模式），补齐工厂注册、eBPF manager 装配、事件解码与输出文件处理；同时按维护者反馈逐轮修正 **符号解析、探针地址配置、模式互斥、keylog/pcapng 行为**，并沉淀为可复用的重构指南/Playbook。另将通用 **PCAPNG 写入能力**下沉到共享 handlers 目录，供后续探针复用。

### 关键改动
- **GoTLS probe 架构对齐（follow bash pattern）**
  - `Config` 继承 `BaseConfig`，`NewProbe` 初始化 `BaseProbe`
  - 完整实现 `Initialize/Start`（含 eBPF manager 创建、options 组装、日志与类型断言）
  - 实现 `EventDecoder`：`Decode()` / `GetDecoder()`
  - `TLSDataEvent` / `MasterSecretEvent` 对齐并实现 `domain.Event`
  - 测试迁移并保持通过（含 race）

- **按维护者反馈的关键修复/增强**
  - 回滚误改的 `kern/` 目录格式化变更
  - CLI 配置与类型修正（移除不必要 cast、补齐 `config.IConfig` 适配器方法）
  - 完成 `register.go` 工厂注册（对标 bash pattern）
  - **多 read uretprobe 支持**：引入 `ReadTlsAddrs[]` 与 `GoTlsWriteAddr`，支持多 offset 挂载与回退逻辑
  - **Go 符号表解析重写**：使用 `debug/gosym` + `debug/buildinfo` 解析 `gopclntab`，并新增覆盖 Go binary/非 Go binary/非法路径的单测
  - **probe 地址配置修正**：统一使用 `UAddress`（绝对地址）替代 `UprobeOffset`
  - **补齐 keylog 与 pcap/TC 支持**
    - 新增 `GoTlsMasterSecretAddr`，定位 `crypto/tls.(*Config).writeKeyLog`
    - pcap 模式加入 TC probes（ingress/egress）及所需 maps，并增加 ifname 校验

- **模式逻辑严格互斥并对齐历史行为**
  - TEXT：仅 TLS 明文抓取相关 probes + events map
  - KEYLOG：仅 `writeKeyLog`（产出密钥）+ mastersecret map，不加载明文 probes
  - PCAP（含 pcapng）：TC 抓包 + `writeKeyLog` 产钥；密钥可写入 PCAPNG DSB 或可选 keylog 文件
  - 规范化 `-m pcap/pcapng`：内部统一视为 `pcap`

- **输出能力补齐**
  - 实现 master secret 写入 **NSS SSLKEYLOGFILE** 格式（Wireshark 可用），覆盖 keylog 与 pcap 模式
  - 引入 `PcapWriter`（基于定制 gopacket fork 的 `pcapgo.NgWriter`）实现标准 PCAPNG 写入与 **DSB（Decryption Secrets Block）** 写钥
  - `PcapWriter` 下沉至 `internal/probe/base/handlers/` 供 openssl/gnutls 等后续复用
  - 按反馈完善 Flush/Close、接口名标识为 `"eCapture(旁观者)"`，并修复相关语法/签名问题

- **文档沉淀**
  - 增加 GoTLS 重构总结
  - 增加可复用的重构指南与更完整的 **Probe Refactoring Playbook**（流程、迭代记录、坑位与检查清单、测试策略、成功指标）

### 结果
- GoTLS 探针在新架构下实现更一致的生命周期、解码与输出能力，keylog/pcapng 行为与历史实现对齐
- 通用 PCAPNG 写入与 DSB 写钥能力完成抽象复用
- 全套测试通过（含 race detector，文中阶段性为 20 tests）。

---------

+++ cli/cmd/gotls.go
@@ -57,8 +57,8 @@ func goTLSCommandFunc(command *cobra.Command, args []string) error {
 		gotlsConfig.PcapFilter = strings.Join(args, " ")
 	}
 
-	// Set global config (note: gotls Config doesn't extend BaseConfig)
-	gotlsConfig.Pid = uint32(globalConf.Pid)
+	// Set global config from BaseConfig
+	gotlsConfig.Pid = globalConf.Pid
 
 	// Run probe using the common entry point
 	return runProbe(factory.ProbeTypeGoTLS, gotlsConfig)
